requirements.txt
google-generativeai
praw
python-dotenv
configparser